Scripting Breeze and TailwindCSS into the system, login draft
The system now uses Tailwind and Breeze.

So, the main view layout is app, one of the pages that extends it is dashboars and inside dashboard, the home is being displayed. Whenever I was trying to do otherwise, the styles were not displaying at all.

The website now has a login feature that was set up alongside with breeze extension.